requirements.txt
odps
datetime
psycopg2-binary
xlsxwriter
scipy
plotly
requests
openpyxl
toml
pygwalker
streamlit
python-dotenv
